backend/requirements.txt
asgiref==3.10.0
coverage==7.10.7
Django==5.2.7
django-cors-headers==4.9.0
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
mccabe==0.7.0
phonenumberslite==9.0.16
pillow==11.3.0
psycopg2==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
pytest==8.3.3
pytest-django==4.9.0


